+++ ios/mapbox_maps_flutter/Sources/mapbox_maps_flutter/Classes/Extensions.swift
@@ -348,6 +348,20 @@ extension LayerPosition {
     }
 }
 
+extension ImportPosition {
+    func toImportPosition() -> MapboxMaps.ImportPosition {
+        var position = MapboxMaps.ImportPosition.default
+        if self.above != nil {
+            position = MapboxMaps.ImportPosition.above(self.above!)
+        } else if self.below != nil {
+            position = MapboxMaps.ImportPosition.below(self.below!)
+        } else if self.at != nil {
+            position = MapboxMaps.ImportPosition.at(Int(at!))
+        }
+        return position
+    }
+}
+
 extension TransitionOptions {
     func toTransitionOptions() -> MapboxMaps.TransitionOptions {
         return MapboxMaps.TransitionOptions(